+++ temba/channels/types/blackmyna/type.py
@@ -1,7 +1,7 @@
 from django.utils.translation import ugettext_lazy as _
 
 from temba.channels.views import AuthenticatedExternalClaimView
-from temba.contacts.models import TEL_SCHEME
+from temba.contacts.models import URN
 
 from ...models import ChannelType
 
@@ -18,19 +18,17 @@ class BlackmynaType(ChannelType):
 
     name = "Blackmyna"
 
-    claim_blurb = _(
-        """Easily add a two way number you have configured with <a href="http://blackmyna.com">Blackmyna</a> using their APIs."""
-    )
+    claim_blurb = _("Easily add a two way number you have configured with %(link)s using their APIs.") % {
+        "link": '<a href="http://blackmyna.com">Blackmyna</a>'
+    }
     claim_view = AuthenticatedExternalClaimView
 
-    schemes = [TEL_SCHEME]
+    schemes = [URN.TEL_SCHEME]
     max_length = 1600
     attachment_support = False
 
     configuration_blurb = _(
-        """
-        To finish configuring your Blackmyna connection you'll need to notify Blackmyna of the following URLs.
-        """
+        "To finish configuring your Blackmyna connection you'll need to notify Blackmyna of the following URLs."
     )
 
     configuration_urls = (