+++ temba/channels/types/chikka/type.py
@@ -1,7 +1,7 @@
 from django.utils.translation import ugettext_lazy as _
 
 from temba.channels.types.chikka.views import ClaimView
-from temba.contacts.models import TEL_SCHEME
+from temba.contacts.models import URN
 
 from ...models import ChannelType
 
@@ -19,19 +19,18 @@ class ChikkaType(ChannelType):
     name = "Chikka"
 
     claim_blurb = _(
-        """If you are based in the Phillipines, you can integrate with Chikka to send
-                       and receive messages on your shortcode."""
+        "If you are based in the Phillipines, you can integrate with Chikka to send and receive "
+        "messages on your shortcode."
     )
     claim_view = ClaimView
 
-    schemes = [TEL_SCHEME]
+    schemes = [URN.TEL_SCHEME]
     max_length = 160
     attachment_support = False
 
     configuration_blurb = _(
-        """
-        To finish configuring your Chikka connection you need to set the following URLs in your Chikka account API settings.
-        """
+        "To finish configuring your Chikka connection you need to set the following URLs in your "
+        "Chikka account API settings."
     )
 
     configuration_urls = (