+++ temba/channels/types/chikka/views.py
@@ -2,12 +2,16 @@ from django import forms
 from django.utils.translation import ugettext_lazy as _
 
 from temba.channels.views import ALL_COUNTRIES, AuthenticatedExternalClaimView, ClaimViewMixin
+from temba.utils.fields import SelectWidget
 
 
 class ClaimView(AuthenticatedExternalClaimView):
     class ChikkaForm(ClaimViewMixin.Form):
         country = forms.ChoiceField(
-            choices=ALL_COUNTRIES, label=_("Country"), help_text=_("The country this phone number is used in")
+            choices=ALL_COUNTRIES,
+            widget=SelectWidget(attrs={"searchable": True}),
+            label=_("Country"),
+            help_text=_("The country this phone number is used in"),
         )
         number = forms.CharField(
             max_length=14, min_length=4, label=_("Number"), help_text=_("The short code you are connecting.")