+++ temba/channels/types/clickmobile/type.py
@@ -1,7 +1,7 @@
 from django.utils.translation import ugettext_lazy as _
 
 from temba.channels.types.clickmobile.views import ClaimView
-from temba.contacts.models import TEL_SCHEME
+from temba.contacts.models import URN
 
 from ...models import ChannelType
 
@@ -20,17 +20,18 @@ class ClickMobileType(ChannelType):
     courier_url = r"^cm/(?P<uuid>[a-z0-9\-]+)/(?P<action>receive)$"
 
     claim_blurb = _(
-        """If you are based in Malawi or Ghana you can purchase a number from %(link)s and connect it in a few simple steps."""
+        "If you are based in Malawi or Ghana you can purchase a number from %(link)s and connect it in a few simple "
+        "steps."
     ) % {"link": '<a href="https://www.click-mobile.com/">Click Mobile</a>'}
 
     claim_view = ClaimView
 
-    schemes = [TEL_SCHEME]
+    schemes = [URN.TEL_SCHEME]
     max_length = 459
     attachment_support = False
 
     configuration_blurb = _(
-        """To finish configuring your channel you need to configure Click Mobile to send new messages to the URL below."""
+        "To finish configuring your channel you need to configure Click Mobile to send new messages to the URL below."
     )
 
     configuration_urls = (