+++ temba/channels/types/dmark/type.py
@@ -1,7 +1,7 @@
 from django.utils.translation import ugettext_lazy as _
 
 from temba.channels.types.dmark.views import ClaimView
-from temba.contacts.models import TEL_SCHEME
+from temba.contacts.models import URN
 
 from ...models import ChannelType
 
@@ -20,20 +20,17 @@ class DMarkType(ChannelType):
     courier_url = r"^dk/(?P<uuid>[a-z0-9\-]+)/(?P<action>receive|status)$"
 
     claim_blurb = _(
-        """If you are based in Uganda or DRC you can purchase a short
-    code from <a href="http://dmarkmobile.com/">DMark Mobile</a> and connect it
-    in a few simple steps."""
-    )
+        "If you are based in Uganda or DRC you can purchase a short code from %(link)s and connect it in a few simple "
+        "steps."
+    ) % {"link": '<a href="http://dmarkmobile.com/">DMark Mobile</a>'}
     claim_view = ClaimView
 
-    schemes = [TEL_SCHEME]
+    schemes = [URN.TEL_SCHEME]
     max_length = 459
     attachment_support = False
 
     configuration_blurb = _(
-        """
-        To finish configuring your DMark channel you need to set DMark to send MO messages to the URL below.
-        """
+        "To finish configuring your DMark channel you need to set DMark to send MO messages to the URL below."
     )
 
     configuration_urls = (