+++ temba/channels/types/facebookapp/type.py
@@ -3,7 +3,7 @@ import requests
 from django.conf.urls import url
 from django.utils.translation import ugettext_lazy as _
 
-from temba.contacts.models import FACEBOOK_SCHEME
+from temba.contacts.models import URN
 
 from ...models import Channel, ChannelType
 from .views import ClaimView, RefreshToken
@@ -27,12 +27,12 @@ class FacebookAppType(ChannelType):
     show_config_page = False
 
     claim_blurb = _(
-        """Add a <a href="http://facebook.com">Facebook</a> bot to send and receive messages on behalf
-    of one of your Facebook pages for free. You will need to connect your page by logging into your Facebook and checking the Facebook page to connect"""
-    )
+        "Add a %(link)s bot to send and receive messages on behalf of one of your Facebook pages for free. You will "
+        "need to connect your page by logging into your Facebook and checking the Facebook page to connect."
+    ) % {"link": '<a href="http://facebook.com">Facebook</a>'}
     claim_view = ClaimView
 
-    schemes = [FACEBOOK_SCHEME]
+    schemes = [URN.FACEBOOK_SCHEME]
     max_length = 2000
     attachment_support = True
     free_sending = True