+++ temba/channels/types/m3tech/type.py
@@ -1,7 +1,7 @@
 from django.utils.translation import ugettext_lazy as _
 
 from temba.channels.views import AuthenticatedExternalClaimView
-from temba.contacts.models import TEL_SCHEME
+from temba.contacts.models import URN
 
 from ...models import ChannelType
 
@@ -18,19 +18,17 @@ class M3TechType(ChannelType):
 
     name = "M3 Tech"
 
-    claim_blurb = _(
-        """Easily add a two way number you have configured with <a href="http://m3techservice.com">M3 Tech</a> using their APIs."""
-    )
+    claim_blurb = _("Easily add a two way number you have configured with %(link)s using their APIs.") % {
+        "link": '<a href="http://m3techservice.com">M3 Tech</a>'
+    }
     claim_view = AuthenticatedExternalClaimView
 
-    schemes = [TEL_SCHEME]
+    schemes = [URN.TEL_SCHEME]
     max_length = 160
     attachment_support = False
 
     configuration_blurb = _(
-        """
-        To finish configuring your connection you'll need to notify M3Tech of the following callback URLs:
-        """
+        "To finish configuring your connection you'll need to notify M3Tech of the following callback URLs."
     )
 
     configuration_urls = (