+++ temba/channels/types/mtarget/type.py
@@ -1,6 +1,6 @@
 from django.utils.translation import ugettext_lazy as _
 
-from temba.contacts.models import TEL_SCHEME
+from temba.contacts.models import URN
 
 from ...models import ChannelType
 from .views import ClaimView
@@ -22,22 +22,17 @@ class MtargetType(ChannelType):
     available_timezones = ["Africa/Kigali", "Africa/Yaoundé", "Africa/Porto-Novo", "Africa/Kinshasa", "Europe/Paris"]
     recommended_timezones = ["Africa/Kigali", "Africa/Yaoundé", "Africa/Porto-Novo", "Africa/Kinshasa", "Europe/Paris"]
 
-    schemes = [TEL_SCHEME]
+    schemes = [URN.TEL_SCHEME]
     max_length = 765
     attachment_support = False
 
     claim_view = ClaimView
-    claim_blurb = _(
-        """
-        If you have an <a href="https://www.mtarget.fr/">Mtarget</a> account,
-        you can quickly connect it using their APIs.
-        """
-    )
+    claim_blurb = _("If you have an %(link)s account, you can quickly connect it using their APIs.") % {
+        "link": '<a href="https://www.mtarget.fr/">Mtarget</a>'
+    }
 
     configuration_blurb = _(
-        """
-        To finish connecting your channel, you need to have Mtarget configure the URLs below for your Service ID.
-        """
+        "To finish connecting your channel, you need to have Mtarget configure the URLs below for your Service ID."
     )
 
     configuration_urls = (