+++ temba/channels/types/twilio_messaging_service/views.py
@@ -7,6 +7,7 @@ from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _
 
 from temba.orgs.models import Org
+from temba.utils.fields import SelectWidget
 
 from ...models import Channel
 from ...views import TWILIO_SUPPORTED_COUNTRIES, ClaimViewMixin
@@ -14,7 +15,9 @@ from ...views import TWILIO_SUPPORTED_COUNTRIES, ClaimViewMixin
 
 class ClaimView(ClaimViewMixin, SmartFormView):
     class TwilioMessagingServiceForm(ClaimViewMixin.Form):
-        country = forms.ChoiceField(choices=TWILIO_SUPPORTED_COUNTRIES)
+        country = forms.ChoiceField(
+            choices=TWILIO_SUPPORTED_COUNTRIES, widget=SelectWidget(attrs={"searchable": True}),
+        )
         messaging_service_sid = forms.CharField(
             label=_("Messaging Service SID"), help_text=_("The Twilio Messaging Service SID")
         )
@@ -46,9 +49,6 @@ class ClaimView(ClaimViewMixin, SmartFormView):
         user = self.request.user
         org = user.get_org()
 
-        if not org:  # pragma: no cover
-            raise Exception(_("No org for this user, cannot claim"))
-
         data = form.cleaned_data
 
         org_config = org.config
@@ -60,7 +60,13 @@ class ClaimView(ClaimViewMixin, SmartFormView):
         }
 
         self.object = Channel.create(
-            org, user, data["country"], "TMS", name=data["messaging_service_sid"], address=None, config=config
+            org,
+            user,
+            data["country"],
+            self.channel_type,
+            name=data["messaging_service_sid"],
+            address=None,
+            config=config,
         )
 
         return super().form_valid(form)