+++ temba/channels/types/whatsapp/type.py
@@ -7,7 +7,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from temba.channels.models import Channel
 from temba.channels.types.whatsapp.views import ClaimView, RefreshView, SyncLogsView, TemplatesView
-from temba.contacts.models import WHATSAPP_SCHEME
+from temba.contacts.models import URN
 from temba.templates.models import TemplateTranslation
 
 from ...models import ChannelType
@@ -114,20 +114,15 @@ class WhatsAppType(ChannelType):
     name = "WhatsApp"
     icon = "icon-whatsapp"
 
-    claim_blurb = _(
-        """If you have an enterprise WhatsApp account, you can connect it to communicate with your contacts"""
-    )
+    claim_blurb = _("If you have an enterprise WhatsApp account, you can connect it to communicate with your contacts")
     claim_view = ClaimView
 
-    schemes = [WHATSAPP_SCHEME]
+    schemes = [URN.WHATSAPP_SCHEME]
     max_length = 4096
     attachment_support = True
 
     def is_available_to(self, user):
         return user.groups.filter(name="Beta")
-
-    def send(self, channel, msg, text):  # pragma: no cover
-        raise Exception("Sending WhatsApp messages is only possible via Courier")
 
     def get_urls(self):
         return [